requirements.txt
PySide6>=6.6
SQLAlchemy>=2.0
mysql-connector-python>=9.0
python-dotenv>=1.0
loguru>=0.7
reportlab>=4.2
